requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
Flask-Caching==2.1.0
Flask-Login==0.6.3
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
email-validator>=2.1.1
Werkzeug==3.0.1
pymongo==4.6.1
